Move VisualSelection to another package
